requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pytemperature==1.1
randfacts==0.15.0
discord-ext-menus==1.1
requests==2.27.1
soupsieve==2.3.2.post1
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
buttons==0.1.9
pymongo==3.12.2
dnspython==2.2.0


